+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Implement `bldb` and update `book_template` (aka `btmpl`) functionalities:
     * ... test all, upd site, publish it  and mk package + publish it
     * ... create a classmethod rhet return MY_BOOKS_ROOT as root (path) of my_books/ directory
     * ... upd `/api/prvb/` ro use this method instrad of function getBook
-    * ... move actual function `getBook` as method of class myBook
+    * c) move actual function `getBook` as method of class myBook
     * b) add class var `db` to have web app initialized database objec6
     * a) boklib: create myBook.MyBook class that manage my books. At init get a book code, Flak application that will use object and set them as instance variables
 * [0.8a11] all scripts moved from booklab/ in booklab/scrpts/ directory